Refactoring the tests
